Make logout button styling local and reorganize folders
